client/requirements.txt
python-dotenv==0.21.0
beautifulsoup4==4.12.2
urllib3==1.26.15